fix: 일정 수정 오류 픽스, Map 불필요한 렌더링 제거, 스타일 수정

+++ src/components/todo/CheckList.js
@@ -17,7 +17,7 @@ const CheckList = ({ state, idSub, sub, subList, setSubList, idCheck, check, dis
       }
       return sub;
     });
-    if (state) {
+    if (state && check.idCheck) {
       try {
         await axios.delete(`/api/todo/check/${idCheck}`);
       } catch (error) {
